Replace parser with a 'gocc'-based BNF grammar
Performance issues were found with the existing parser and
evaluation system especially in tight loop conditions.  The
current approach of a custom lexer and reverse Polish-style
evaluation of the result is compact and an appropriate
algorithm for the use case, but unfortunately needs a lot of
work to become performant.

Instead it's possible to drop-in wholesale a gocc-generated,
BNF specified parser.  This parser can generate an AST which
can evaluate in an easier to follow fashion, likely with a
lot better performance out of the box.  Start with a
degenerate grammar that only parses 'null'.