Regular Expression Operations
Support the regexp-related operators in sql_operators.go.
The grammar forces the use of a literal regular expression
to match against instead of allowing a dynamic expression,
and stores the compiled regular expression in the AST.  This
is a little restrictive but steers use towards the best use
case for regexps, especially DFA-based as in go: a single
compiled DFA matches many inputs.

There is also some perl-style stringification of values if
they happen to end up as a match condition, which is close
to what JavaScript does as well I guess.